Debug integer overflow

+++ pyscf/lib/pbc/fill_ints.c
@@ -450,7 +450,7 @@ static void _nr3c_fill_k(int (*intor)(), void (*fsort)(),
                 dijmk = dijmc * nkpts;
                 bufk_i = bufk_r + dijmk;
                 bufL   = bufk_i + dijmk;
-                cache  = bufL   + nimgs * dijmc;
+                cache  = bufL   + ((size_t)nimgs) * dijmc;
                 for (i = 0; i < dijmk*OF_CMPLX; i++) {
                         bufk_r[i] = 0;
                 }
